chore(testing): extracted BasePickerResults and stubbed out a spec
