+++ src/elements/picker/Picker.module.ts
@@ -3,8 +3,8 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 // APP
-import { NovoListModule } from './../list/List.module';
-import { NovoLoadingModule } from './../loading/Loading.module';
+import { NovoListModule } from '../list/List.module';
+import { NovoLoadingModule } from '../loading/Loading.module';
 import { NovoPickerElement, NovoPickerContainer } from './Picker';
 import { PickerResults } from './extras/picker-results/PickerResults';
 import { EntityPickerResult, EntityPickerResults } from './extras/entity-picker-results/EntityPickerResults';
