+++ src/elements/picker/extras/picker-results/PickerResults.spec.ts
@@ -1,29 +1,29 @@
-// TODO
-// import { Component } from '@angular/core';
-//
-// import { PickerResults } from './PickerResults';
-// import { testComponent, grabComponent } from './../../../../testing/TestHelpers';
-// import { NOVO_ELEMENTS_LABELS_PROVIDERS } from './../../../../novo-elements';
-//
-// @Component({
-//     selector: 'test-cmp',
-//     directives: [PickerResults],
-//     template: '<picker-results></picker-results>'
-// })
-// class TestCmp {
-//     constructor() {
-//         this.options = ['Alabama', 'Alaska', 'Arizona', 'Arkansas', 'California', 'Colorado', 'Connecticut', 'Delaware', 'Florida', 'Georgia', 'Hawaii', 'Idaho', 'Illinois', 'Indiana', 'Iowa', 'Kansas', 'Kentucky', 'Louisiana', 'Maine', 'Maryland', 'Massachusetts', 'Michigan', 'Minnesota', 'Mississippi', 'Missouri', 'Montana', 'Nebraska', 'Nevada', 'New Hampshire', 'New Jersey', 'New Mexico', 'New York', 'North Dakota', 'North Carolina', 'Ohio', 'Oklahoma', 'Oregon', 'Pennsylvania', 'Rhode Island', 'South Carolina', 'South Dakota', 'Tennessee', 'Texas', 'Utah', 'Vermont', 'Virginia', 'Washington', 'West Virginia', 'Wisconsin', 'Wyoming'];
-//     }
-// }
-//
-// describe('Element: PickerResults', () => {
-//     beforeEachProviders(() => [NOVO_ELEMENTS_LABELS_PROVIDERS]);
-//
-//     it('should initialize correctly', testComponent(TestCmp, (fixture) => {
-//         const { instance, element, testComponentInstance, testComponentElement } = grabComponent(fixture, PickerResults);
-//         expect(instance).toBeTruthy();
-//         expect(element).toBeTruthy();
-//         expect(testComponentInstance).toBeTruthy();
-//         expect(testComponentElement).toBeTruthy();
-//     }));
-// });
+// NG2
+import { TestBed, async } from '@angular/core/testing';
+// App
+import { PickerResults } from './PickerResults';
+import { NovoLabelService } from '../../../../services/novo-label-service';
+import { NovoLoadingElement } from '../../../loading/Loading';
+
+describe('Elements: PickerResults', () => {
+    let fixture;
+    let component;
+
+    beforeEach(async(() => {
+        TestBed.configureTestingModule({
+            declarations: [
+                PickerResults,
+                NovoLoadingElement
+            ],
+            providers: [
+                { provide: NovoLabelService, useClass: NovoLabelService }
+            ]
+        }).compileComponents();
+        fixture = TestBed.createComponent(PickerResults);
+        component = fixture.debugElement.componentInstance;
+    }));
+
+    it('should be defined.', () => {
+        expect(component).toBeDefined();
+    });
+});
